add #凸状況の履歴
