+++ config/settings/channel.yaml
@@ -132,7 +132,15 @@ NOT_EMOJI_CHANNEL:
   - 5ボス-凸宣言
   - 凸報告
   - 持越状況
-  - 凸可能時間
   - 凸予定
   - 凸状況
-  - 凸状況履歴
+  - bot操作-履歴
+  - 1ボス-凸宣言-履歴
+  - 2ボス-凸宣言-履歴
+  - 3ボス-凸宣言-履歴
+  - 4ボス-凸宣言-履歴
+  - 5ボス-凸宣言-履歴
+  - 持越状況-履歴
+  - 凸予定-履歴
+  - 凸状況-履歴
+  - ボス状況-履歴
